moved sector components into their own submodule
