+++ src/components/sector/mod.rs
@@ -1,0 +1,6 @@
+mod in_sector;
+mod sector_asteroid;
+mod sector_component;
+mod sector_star;
+
+pub use {in_sector::*, sector_asteroid::*, sector_component::*, sector_star::*};
